Tests for select eqn2mml endpoints
